normalize daily prices and enable daily k line on front end
