+++ tanglism-web/static/index.html
@@ -19,6 +19,7 @@
                 <option>1m</option>
                 <option>5m</option>
                 <option selected>30m</option>
+                <option>1d</option>
             </select>
             <!-- <input type="checkbox" id="k_check" value="k" checked/>
             <label for="k_check">K线</label> -->
